requirements.txt
matplotlib==3.10.5
numba==0.60.0
numpy==2.3.2
opencv_python==4.10.0.84
pandas==2.3.1
Pillow==11.3.0
protobuf==6.31.1
pycocotools==2.0.8
PyYAML==6.0.2
retrying==1.3.4
sam_2==1.0
scikit_learn==1.7.1
torch==2.6.0
transformers==4.54.1
ultralytics==8.3.159
